Move WMTSLoader until math utility sharing is figured out
