+++ src/core/plugins/index.d.ts
@@ -1,3 +1,2 @@
 export * from './ImplicitTilingPlugin.js';
 export * from './EnforceNonZeroErrorPlugin.js';
-export * from './loaders/WMTSCapabilitiesLoader.js';
